4th attempt

+++ src/app/page.tsx
@@ -105,7 +105,8 @@ export default function VoiceAssistant() {
         const detectSilence = () => {
           if (!analyser.current || !dataArray.current) return
 
-          analyser.current.getByteFrequencyData(dataArray.current as Uint8Array)
+          // Fix: Remove the type assertion and handle the Uint8Array properly
+          analyser.current.getByteFrequencyData(dataArray.current)
           const average = dataArray.current.reduce((acc, val) => acc + val, 0) / dataArray.current.length
           setAudioLevel(average)
